🔧 Jules05: Resolve cross-agent conflict in Risk Management and Model interfaces
- Harmonized Risk Management API by merging RiskEngine logic into RiskManager and AuditedRiskManager.
- Decommissioned redundant src/trading/risk_engine.py.
- Standardized Model.predict interface to consistently forward context via **kwargs.
- Updated src/trading/__init__.py with correct exports.
- Appended findings and resolutions to docs/integration/CONFLICT_LOG.md.
- Added verification tests in tests/test_risk_manager_harmonized.py.
