+++ src/trading/__init__.py
@@ -3,6 +3,7 @@
 from __future__ import annotations
 
 from src.core.schemas import TradeSignal
+from src.trading.audited_risk_manager import AuditedRiskManager
 from src.trading.backtester import BacktestEngine, PerformanceReport
 from src.trading.capital_allocator import (
     AllocationRequest,
@@ -12,18 +13,18 @@ from src.trading.capital_allocator import (
     StrategyConfig,
 )
 from src.trading.mt5_connector import MT5Connector
-from src.trading.risk_engine import RiskDecision, RiskEngine
+from src.trading.risk_manager import RiskManager
 
 __all__ = [
     "AllocationRequest",
     "AllocationResult",
+    "AuditedRiskManager",
     "BacktestEngine",
     "CapitalAllocator",
     "MT5Connector",
     "PerformanceReport",
     "RejectionCode",
-    "RiskDecision",
-    "RiskEngine",
+    "RiskManager",
     "StrategyConfig",
     "TradeSignal",
 ]
