+++ src/trading/audited_risk_manager.py
@@ -9,7 +9,9 @@ License: MIT
 from __future__ import annotations
 
 import logging
-from typing import Optional
+from typing import Any, Dict, List, Optional
+
+import pandas as pd
 
 from src.core.audit_log import get_audit_logger
 from src.core.schemas import TradeSignal
@@ -29,6 +31,8 @@ class AuditedRiskManager(RiskManager):
         signal: TradeSignal,
         signal_id: Optional[int] = None,
         model_health: Optional[dict] = None,
+        market_data: Optional[pd.DataFrame] = None,
+        open_positions: Optional[List[Dict[str, Any]]] = None,
     ) -> bool:
         """
         Run the full 8-layer risk filter cascade.
@@ -39,6 +43,12 @@ class AuditedRiskManager(RiskManager):
             "circuit_breaker": self._check_circuit_breaker(),
             "daily_loss": self._check_daily_loss(),
             "max_positions": self._check_max_positions(),
+            "directional_exposure": self._check_directional_exposure(signal, open_positions)
+            if open_positions is not None
+            else True,
+            "total_notional": self._check_total_notional(signal, open_positions, market_data)
+            if open_positions is not None and market_data is not None
+            else True,
             "symbol_allocation": self._check_symbol_allocation(signal.symbol),
             "min_confidence": self._check_minimum_confidence(signal.confidence),
             "risk_reward": self._check_risk_reward(signal),
