+++ tests/test_risk_manager_harmonized.py
@@ -71,17 +71,15 @@ def test_drawdown_breaker(risk_manager, buy_signal, market_data):
     risk_manager.peak_equity = 20000.0
     risk_manager.balance = 10000.0  # 50% drawdown
 
-    decision = risk_manager.validate_signal(buy_signal, market_data, [])
-    assert not decision.is_approved
-    assert "drawdown" in decision.reason.lower()
+    approved = risk_manager.approve(buy_signal, market_data=market_data, open_positions=[])
+    assert not approved
 
 def test_daily_loss_limit(risk_manager, buy_signal, market_data):
     risk_manager.daily.peak_equity = 10000.0
     risk_manager.daily.realised_pnl = -600.0  # 6% loss
 
-    decision = risk_manager.validate_signal(buy_signal, market_data, [])
-    assert not decision.is_approved
-    assert "daily loss" in decision.reason.lower()
+    approved = risk_manager.approve(buy_signal, market_data=market_data, open_positions=[])
+    assert not approved
 
 def test_max_positions(risk_manager, buy_signal, market_data):
     open_positions = [
@@ -90,9 +88,8 @@ def test_max_positions(risk_manager, buy_signal, market_data):
         {"ticket": 3, "symbol": "XAUUSD", "volume": 0.1, "type": 0},
     ]
 
-    decision = risk_manager.validate_signal(buy_signal, market_data, open_positions)
-    assert not decision.is_approved
-    assert "max concurrent positions" in decision.reason.lower()
+    approved = risk_manager.approve(buy_signal, market_data=market_data, open_positions=open_positions)
+    assert not approved
 
 def test_directional_exposure(risk_manager, buy_signal, market_data):
     # Max single direction is 30% of 10000 = 3000
@@ -102,24 +99,21 @@ def test_directional_exposure(risk_manager, buy_signal, market_data):
         {"ticket": 1, "symbol": "XAUUSD", "volume": 0.13, "type": 0},
     ]
 
-    decision = risk_manager.validate_signal(buy_signal, market_data, open_positions)
-    assert not decision.is_approved
-    assert "directional exposure" in decision.reason.lower()
+    approved = risk_manager.approve(buy_signal, market_data=market_data, open_positions=open_positions)
+    assert not approved
 
 def test_atr_position_sizing(risk_manager, market_data):
     # Normal volatility
     market_data["atr"] = 1.0
-    size = risk_manager.size_position("XAUUSD", market_data)
+    size = risk_manager.size_position_atr("XAUUSD", market_data)
     assert size > 0
 
     # Extreme volatility
     market_data.loc[market_data.index[-1], "atr"] = 4.0
     # avg_atr remains approx 1.0. ratio = 4.0 > 3.0 (extreme threshold)
-    size = risk_manager.size_position("XAUUSD", market_data)
+    size = risk_manager.size_position_atr("XAUUSD", market_data)
     assert size == 0.0
 
 def test_full_approval(risk_manager, buy_signal, market_data):
-    decision = risk_manager.validate_signal(buy_signal, market_data, [])
-    assert decision.is_approved
-    assert decision.reason == "Approved"
-    assert decision.adjusted_lot_size > 0
+    approved = risk_manager.approve(buy_signal, market_data=market_data, open_positions=[])
+    assert approved
